Adding current gesture to screen
