+++ app/templates/index.html
@@ -6,5 +6,15 @@
 <body>
     <h1>Video Stream</h1>
     <img src="{{ url_for('video_feed') }}" width="640" height="480">
+    <div id="gesture">Current Gesture: No gesture detected yet</div>
+    <script>
+    setInterval(function() {
+        fetch('/current_gesture')
+            .then(response => response.json())
+            .then(data => {
+                document.getElementById('gesture').innerHTML = 'Current Gesture: ' + data.gesture;
+            });
+    }, 1000);
+    </script>
 </body>
 </html>
